Extract product inventory cockpit
